% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/example3.R
\name{example1}
\alias{example1}
\alias{example2}
\alias{example3}
\title{spdlite Example using a sink for R}
\usage{
example1()

example2()

example3()
}
\value{
None
}
\description{
A simple example invoking a logger from C++. It can interact with logging from R.
}
