% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{log_trace}
\alias{log_trace}
\alias{log_debug}
\alias{log_info}
\alias{log_warn}
\alias{log_error}
\alias{log_critical}
\alias{set_level}
\alias{get_level}
\alias{set_name}
\alias{get_name}
\alias{set_precision}
\alias{show_thread_id}
\alias{show_date}
\alias{show_utc}
\alias{set_format}
\title{Convenience Wrappers for 'spdlite' Logging}
\usage{
log_trace(s, ...)

log_debug(s, ...)

log_info(s, ...)

log_warn(s, ...)

log_error(s, ...)

log_critical(s, ...)

set_level(level)

get_level()

set_name(s)

get_name()

set_precision(precision)

show_thread_id(show_thread_id = TRUE)

show_date(show_date = TRUE)

show_utc(utc = TRUE)

set_format(utc = FALSE, show_date = TRUE, show_thread_id = FALSE,
  precision = "ms")
}
\arguments{
\item{s}{Character value for filename, pattern, level, or logging message}

\item{...}{Supplementary arguments for the logging string}

\item{level}{Character value for the logging level}

\item{precision}{Character value for selected time precision: one of \dQuote{ms}
(the default format), \dQuote{us}, \dQuote{ns} or \dQuote{none}}

\item{show_thread_id}{Boolean flag select display of current thread, default is off}

\item{show_date}{Boolean flag select display of date part of current, default is on}

\item{utc}{Boolean flag select display of current time in UTC rather than local,
default is off}
}
\value{
Nothing is returned from these functions (with the exception of
\code{get_level()}) as they are invoked for their side-effects.
}
\description{
Several short wrappers for functions from 'spdlite' are provided as a convenience.
}
