% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{rspdlite-package}
\alias{rspdlite}
\alias{rspdlite-package}
\title{rspdlite: Lightweight Logging From R and C++}
\description{
\pkg{rspdlite} relies on \pkg{spdlite}, a lightweight header-only C++20
library for logging. This package offers a simple and consistent interface
from both R and C++.
}
\details{
\pkg{spdlite} is the \sQuote{little brother} of \pkg{spdlog} (which we have
available via R packages \pkg{RcppSpdlog} and \pkg{spdl}). \pkg{spdlite} is
on purpose smaller and simpler. It has (by choice) fewer options and
configuration settings keeping the core small and simple.

By using a global \code{inline} instance, each shared library (i.e. typically
each package using logging) is guaranteed to have exactly one instance. So
setting changes such as the logging level affect both the R and C++ side. If
however another R package were to be compiled with the \pkg{spdlite} header,
its instance would be separate as it resides in a different shared library.
Similarly, an ad-hoc compilation via e.g. \code{Rcpp::sourceCpp()} will lead
to a distinct instance for the same reason.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/eddelbuettel/rspdlite}
  \item Report bugs at \url{https://github.com/eddelbuettel/rspdlite/issues}
}

}
\author{
\strong{Maintainer}: Dirk Eddelbuettel \email{edd@debian.org} (0000-0001-6419-907X)

Authors:
\itemize{
  \item Gabi Melman (Author of spdlog)
  \item Victor Zverovic (Author of fmt)
}

}
