% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{trace}
\alias{trace}
\alias{debug}
\alias{info}
\alias{warn}
\alias{error}
\alias{critical}
\alias{set_level}
\alias{get_level}
\title{Convenience Wrappers for 'spdlite' Logging}
\usage{
trace(s, ...)

debug(s, ...)

info(s, ...)

warn(s, ...)

error(s, ...)

critical(s, ...)

set_level(level)

get_level()
}
\arguments{
\item{s}{Character value for filename, pattern, level, or logging message}

\item{...}{Supplementary arguments for the logging string}

\item{level}{Character value for the logging level}
}
\value{
Nothing is returned from these functions (with the exception of
\code{get_level()}) as they are invoked for their side-effects.
}
\description{
Several short wrappers for functions from 'spdlite' are provided as a convenience.
}
